+++ spec/rubocop/cop/style/encoding_spec.rb
@@ -11,14 +11,14 @@ describe RuboCop::Cop::Style::Encoding, :config do
     end
 
     it 'registers no offense when no encoding present but only ASCII ' \
-       'characters', ruby: 1.9 do
+       'characters' do
       inspect_source(cop, ['def foo() end'])
 
       expect(cop.offenses).to be_empty
     end
 
     it 'registers an offense when there is no encoding present but non ' \
-       'ASCII characters', ruby: 1.9 do
+       'ASCII characters' do
       inspect_source(cop, ['def foo() \'ä\' end'])
 
       expect(cop.offenses.size).to eq(1)
@@ -27,7 +27,7 @@ describe RuboCop::Cop::Style::Encoding, :config do
     end
 
     it 'registers an offense when encoding present but only ASCII ' \
-       'characters', ruby: 1.9 do
+       'characters' do
       inspect_source(cop, ['# encoding: utf-8',
                            'def foo() end'])
 
@@ -36,20 +36,20 @@ describe RuboCop::Cop::Style::Encoding, :config do
         ['Unnecessary utf-8 encoding comment.'])
     end
 
-    it 'accepts an empty file', ruby: 1.9 do
+    it 'accepts an empty file' do
       inspect_source(cop, '')
 
       expect(cop.offenses).to be_empty
     end
 
-    it 'accepts encoding on first line', ruby: 1.9 do
+    it 'accepts encoding on first line' do
       inspect_source(cop, ['# encoding: utf-8',
                            'def foo() \'ä\' end'])
 
       expect(cop.offenses).to be_empty
     end
 
-    it 'accepts encoding on second line when shebang present', ruby: 1.9 do
+    it 'accepts encoding on second line when shebang present' do
       inspect_source(cop, ['#!/usr/bin/env ruby',
                            '# encoding: utf-8',
                            'def foo() \'ä\' end'])
@@ -57,7 +57,7 @@ describe RuboCop::Cop::Style::Encoding, :config do
       expect(cop.messages).to be_empty
     end
 
-    it 'books an offense when encoding is in the wrong place', ruby: 1.9 do
+    it 'books an offense when encoding is in the wrong place' do
       inspect_source(cop, ['def foo() \'ä\' end',
                            '# encoding: utf-8'])
 
@@ -66,20 +66,14 @@ describe RuboCop::Cop::Style::Encoding, :config do
         ['Missing utf-8 encoding comment.'])
     end
 
-    it 'does not register an offense on Ruby 2.0', ruby: 2.0 do
-      inspect_source(cop, ['def foo() \'ä\' end'])
-
-      expect(cop.offenses).to be_empty
-    end
-
-    it 'accepts encoding inserted by magic_encoding gem', ruby: 1.9 do
+    it 'accepts encoding inserted by magic_encoding gem' do
       inspect_source(cop, ['# -*- encoding : utf-8 -*-',
                            'def foo() \'ä\' end'])
 
       expect(cop.messages).to be_empty
     end
 
-    it 'accepts vim-style encoding comments', ruby: 1.9 do
+    it 'accepts vim-style encoding comments' do
       inspect_source(cop, ['# vim:fileencoding=utf-8',
                            'def foo() \'ä\' end'])
       expect(cop.messages).to be_empty
@@ -91,7 +85,7 @@ describe RuboCop::Cop::Style::Encoding, :config do
       { 'EnforcedStyle' => 'always' }
     end
 
-    it 'registers an offense when no encoding present', ruby: 1.9 do
+    it 'registers an offense when no encoding present' do
       inspect_source(cop, ['def foo() end'])
 
       expect(cop.offenses.size).to eq(1)
@@ -99,20 +93,20 @@ describe RuboCop::Cop::Style::Encoding, :config do
         ['Missing utf-8 encoding comment.'])
     end
 
-    it 'accepts an empty file', ruby: 1.9 do
+    it 'accepts an empty file' do
       inspect_source(cop, '')
 
       expect(cop.offenses).to be_empty
     end
 
-    it 'accepts encoding on first line', ruby: 1.9 do
+    it 'accepts encoding on first line' do
       inspect_source(cop, ['# encoding: utf-8',
                            'def foo() end'])
 
       expect(cop.offenses).to be_empty
     end
 
-    it 'accepts encoding on second line when shebang present', ruby: 1.9 do
+    it 'accepts encoding on second line when shebang present' do
       inspect_source(cop, ['#!/usr/bin/env ruby',
                            '# encoding: utf-8',
                            'def foo() end'])
@@ -120,7 +114,7 @@ describe RuboCop::Cop::Style::Encoding, :config do
       expect(cop.messages).to be_empty
     end
 
-    it 'books an offense when encoding is in the wrong place', ruby: 1.9 do
+    it 'books an offense when encoding is in the wrong place' do
       inspect_source(cop, ['def foo() end',
                            '# encoding: utf-8'])
 
@@ -129,20 +123,14 @@ describe RuboCop::Cop::Style::Encoding, :config do
         ['Missing utf-8 encoding comment.'])
     end
 
-    it 'does not register an offense on Ruby 2.0', ruby: 2.0 do
-      inspect_source(cop, ['def foo() end'])
-
-      expect(cop.offenses).to be_empty
-    end
-
-    it 'accepts encoding inserted by magic_encoding gem', ruby: 1.9 do
+    it 'accepts encoding inserted by magic_encoding gem' do
       inspect_source(cop, ['# -*- encoding : utf-8 -*-',
                            'def foo() end'])
 
       expect(cop.messages).to be_empty
     end
 
-    it 'accepts vim-style encoding comments', ruby: 1.9 do
+    it 'accepts vim-style encoding comments' do
       inspect_source(cop, ['# vim:fileencoding=utf-8',
                            'def foo() end'])
       expect(cop.messages).to be_empty